Make it easier to reuse templates client for Classic templates
