+++ it/src/main/resources/test-artifact.txt
