+++ v2/streaming-data-generator/pom.xml
@@ -71,12 +71,6 @@
       <version>${truth.version}</version>
       <scope>test</scope>
     </dependency>
-    <dependency>
-      <groupId>com.google.cloud.teleport.v2</groupId>
-      <artifactId>testing</artifactId>
-      <version>1.0-SNAPSHOT</version>
-      <scope>test</scope>
-    </dependency>
   </dependencies>
 
   <build>
