--- v2/testing/src/main/java/com/google/cloud/teleport/v2/testing/artifacts/package-info.java
@@ -1,2 +1,0 @@
-/** Package for working with test artifacts. */
-package com.google.cloud.teleport.v2.testing.artifacts;
