--- v2/testing/src/main/java/com/google/cloud/teleport/v2/testing/dataflow/package-info.java
@@ -1,2 +1,0 @@
-/** Package for managing Dataflow jobs from integration tests. */
-package com.google.cloud.teleport.v2.testing.dataflow;
