--- v2/testing/src/main/java/com/google/cloud/teleport/v2/testing/package-info.java
@@ -1,2 +1,0 @@
-/** Package for general-purpose testing utilities. */
-package com.google.cloud.teleport.v2.testing;
